feat: Implementation and begin of tests

+++ src/services/FileStorage.ts
@@ -1,7 +1,26 @@
 import { existsSync, lstatSync, writeFileSync } from 'fs';
 import { readFile, unlink, writeFile } from 'fs/promises';
 import { isAbsolute } from 'path';
-import type StorageEngine from '../interfaces/StorageEngine';
+import type StorageEngine from '../interfaces/StorageEngine.js';
+
+/**
+ * Provides a factory function to create a file-based storage engine for storing and retrieving JSON data.
+ *
+ * @template Schema - The type of the data schema to be stored.
+ * @param initialData - Optional initial data to populate the file if it does not exist.
+ * @returns A factory function that takes a data path and returns a storage engine.
+ */
+export function FileStorageFactory<Schema extends object>(
+    initialData?: Schema
+) {
+    /**
+     * Creates a file-based storage engine for the specified data path.
+     *
+     * @param dataPath - The absolute path to the file where data will be stored.
+     * @returns A storage engine with methods to write, read, and remove data.
+     */
+    return (dataPath: string) => FileStorage(dataPath, initialData);
+}
 
 /**
  * Provides a file-based storage engine for storing and retrieving JSON data.
@@ -16,7 +35,11 @@ export default function FileStorage<Schema extends object>(
     initialData?: Schema
 ): StorageEngine<Schema> {
     // Validate the provided path
-    if (!path || !isAbsolute(path) || lstatSync(path).isDirectory()) {
+    if (
+        !path ||
+        !isAbsolute(path) ||
+        (existsSync(path) && lstatSync(path).isDirectory())
+    ) {
         throw new Error('Invalid file path');
     }
 
@@ -26,9 +49,7 @@ export default function FileStorage<Schema extends object>(
     }
 
     // Implement the StorageEngine interface methods
-    async function write<LocalSchema extends object = Schema>(
-        value: LocalSchema
-    ): Promise<void> {
+    async function write(value: Schema): Promise<void> {
         await writeFile(path, JSON.stringify(value, null, 2));
     }
 
@@ -36,9 +57,13 @@ export default function FileStorage<Schema extends object>(
      * Reads and parses the JSON data from the file.
      * @returns A promise that resolves with the parsed data.
      */
-    async function read<LocalSchema extends object = Schema>(): Promise<LocalSchema> {
+    async function read(): Promise<Schema> {
+        if (!existsSync(path)) {
+            await write((initialData ?? {}) as Schema);
+        }
+
         const data = await readFile(path, 'utf-8');
-        return JSON.parse(data) as LocalSchema;
+        return JSON.parse(data) as Schema;
     }
 
     /**
@@ -46,7 +71,7 @@ export default function FileStorage<Schema extends object>(
      * @returns A promise that resolves when the file is removed.
      */
     async function remove(): Promise<void> {
-        await unlink(path);
+        if (existsSync(path)) await unlink(path);
     }
 
     return {
